Disable exception in Core LRG test CacheStatementBatchWritingTest
The test
org.eclipse.persistence.testing.tests.feature.CacheStatementBatchWritingTest.CacheStatementBatchWritingTest
is known to be failing with the message
"Exception Description: Statements were not cached correctly".
This error is caused by a bug in the HANA JDBC driver reported as CSN
message 516888 2013.

The exception is temporarily disabled because this is a known bug.

Signed-off-by: Sabine Heider <sabine.heider@sap.com>