Disable ClientServerTestModel in core tests
The core LRG test suite
org.eclipse.persistence.testing.tests.clientserver.ClientServerTestModel
hangs occasionally on HANA. In the HANA Studio one can see that there is
a SQLExecutor thread hanging which is about to execute the statement
DELETE FROM PHONE WHERE ((EMP_ID = ?) AND (TYPE = ?))
Investigations showed that the hanging test is
org.eclipse.persistence.testing.tests.unitofwork.UnitOfWorkConcurrentRevertTest
However, it is not clear if this is the only test causing this lock
situation.

Until the tests have been investigated, the entire test suite is
disabled.
Further analysis is needed.

Signed-off-by: Sabine Heider <sabine.heider@sap.com>